Abbasfromhost

+++ app/Http/Controllers/CareerController.php
@@ -1,13 +1,18 @@
 <?php
+
 namespace App\Http\Controllers;
 
 use App\Models\career;
 use App\Models\careerCategory;
+use App\Models\menu_item;
 use App\Models\province_cities;
 use App\Models\qr_code;
 use App\Models\role_user;
 use App\Models\User;
+use App\Models\menu;
+use App\Models\menu_category;
 use App\Models\order;
+use App\Models\pages;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
@@ -25,19 +30,32 @@ class CareerController extends Controller
         if (!$user) {
             $user = Auth::user();
         }
-        return view('admin.careers.create', ['user' => $user, 'careerCategories' => $careerCategories, 'provinces' => $provinces, 'cities'=>$cities]);
+        return view('admin.careers.create', ['user' => $user, 'careerCategories' => $careerCategories, 'provinces' => $provinces, 'cities' => $cities]);
+    }
+
+    public function createRestaurant(User $user = null, pages $page)
+    {
+
+        $careerCategories = careerCategory::all();
+        if (!$user) {
+            $user = Auth::user();
+        }
+        return view('admin.careers.createRestaurant', ['user' => $user, 'careerCategories' => $careerCategories, 'page' => $page]);
     }
 
     public function store(Request $request)
     {
-        // $user = user::find($request->user_id);
-        // dd($request->all());
+        $user = user::find($request->user_id);
+        $page_id = 0;
         $path = null;
         $bannerPath = null;
-        role_user::create([
-            'user_id'=>$request->user_id,
-            'role_id'=>2
-        ]);
+        $ids = Auth::user()->role->pluck('id')->toArray();
+        if (!in_array(2, $ids)) {
+            role_user::create([
+                'user_id' => $request->user_id,
+                'role_id' => 2
+            ]);
+        }
         if (isset($request->logo)) {
             $name = $request->logo->getClientOriginalName();
             $fullName = Str::uuid() . '_' . $name;
@@ -48,6 +66,22 @@ class CareerController extends Controller
             $fullBannerName = Str::uuid() . '_' . $bannerName;
             $bannerPath = $request->file('banner')->storeAs('files', $fullBannerName, 'public');
         }
+        if (count($user->pages) == 0) {
+            $page_id = pages::insertGetId([
+                'title' => $request->title,
+                'subTitle' => "",
+                'description' => $request->description,
+                'user_id' => $request->user_id,
+                'logo_path' => $path,
+                'cover_path' => $bannerPath,
+            ]);
+        }
+        if (count($user->pages) > 0) {
+            $page_id = $user->pages[0]->id;
+        }
+        if (isset($request->page_id)) {
+            $page_id = $request->page_id;
+        }
         $social_medias = json_encode($request->social_medias);
         $career_id = career::insertGetId([
             'title' => $request->title,
@@ -57,11 +91,16 @@ class CareerController extends Controller
             'social_media' => $social_medias,
             'user_id' => $request->user_id,
             'email' => $request->email,
+            'phone' => isset($request->phone) ? $request->phone : null,
             'description' => $request->description,
             'career_category_id' => $request->careerCategory,
             'banner' => $bannerPath,
-            'qr_count' => $request->qr_count ? $request->qr_count : 0
+            'qr_count' => $request->qr_count ? $request->qr_count : 0,
+            'page_id' => $page_id,
+            'show_in_home'=> $request->show_in_home ? 1 : 0,
+            'active'=> $request->active ? 1 : 0
         ]);
+        $counter = 1;
         for ($i = 0; $i < $request->qr_count; $i++) {
             $random = Str::random(10);
             $link = "famenu.ir/qrcode/$career_id/" . $random;
@@ -73,10 +112,70 @@ class CareerController extends Controller
                 'career_id' => $career_id,
                 'page_path' => 'qrcode/' . $career_id . '/' . $random,
                 'slug' => $random,
+                'description'=>'میز '.$counter,
                 'user_id' => $request->user_id
             ]);
+            $counter++;
+        }
+        if (isset($request->page_id)) {
+//            menu create
+            $menu_id = menu::insertGetId([
+                'title' => 'منو 1',
+                'subtitle' => null,
+                'banner' => null,
+                'career_id' => $career_id,
+                'user_id' => $request->user_id,
+                'created_at' => now(),
+                'updated_at' => now()
+            ]);
+
+//            menu category create
+            $menu_cat_id = menu_category::insertGetId([
+                'title'=>'بدون دسته بندی',
+                'description'=>'آیتم هایی که زیر مجموعه دسته ای نباشند در این دسته قرار میگیرند',
+                'image'=>null,
+                'menu_id'=>$menu_id,
+                'created_at'=>now(),
+                'updated_at'=>now()
+            ]);
+
+//            menu item create
+            menu_item::insertGetId([
+                'title' => 'آیتم 1',
+                'description' => null,
+                'price' => 0,
+                'discount' => 0,
+                'customizable' => 0,
+                'image' => null,
+                'parent_id' => 0,
+                'menu_category_id' => $menu_cat_id,
+                'duration' => 0,
+                'created_at' => now(),
+                'updated_at' => now()
+            ]);
+            $career = career::find($career_id);
+            return response()->json($career);
+//            return to_route('career.showWithMenu', [$career_id]);
         }
         return to_route('career.careers', ['user' => $request->user_id]);
+    }
+
+
+    public function showWithMenu(career $career)
+    {
+        $menus = $career->menus;
+        foreach ($menus as $menu) {
+            foreach ($menu->menu_categories as $category) {
+                foreach ($category->menu_items as $item) {
+                    if($item->discount != 0){
+                        $campare = $item->price - $item->discount;
+                        $x = $campare / $item->price;
+                        $item['percent'] = intval($x * 100);
+                    }
+                }
+            }
+        }
+        return view('admin.careers.showWithMenu', ['career' => $career]);
     }
 
     public function user_careers(user $user = null)
@@ -90,7 +189,11 @@ class CareerController extends Controller
     public function edit(career $career, User $user = null)
     {
         $provinces = province_cities::where('parent', 0)->get();
-        $cities = province_cities::where('parent', $career->province_city->province->id)->get();
+        $cities = province_cities::where('parent', 1)->get();
+        if(isset($career->province_city)){
+            $provinces = province_cities::where('parent', 0)->get();
+            $cities = province_cities::where('parent', $career->province_city->province->id)->get();
+        }
         if (!$user) {
             $user = Auth::user();
         }
@@ -128,11 +231,15 @@ class CareerController extends Controller
         $career->address = $request->address;
         $career->description = $request->description;
         $career->email = $request->email;
+        $career->phone = $request->phone;
         $career->user_id = $request->user_id;
+        $career->show_in_home = $request->show_in_home ? 1 : 0;
+        $career->active= $request->active ? 1 : 0;
         $career->career_category_id = $request->careerCategory;
         if ($request->qr_count) {
-            if ((int) $request->qr_count > $career->qr_count) {
-                $qr_count = (int) $request->qr_count - $career->qr_count;
+            if ((int)$request->qr_count > $career->qr_count) {
+                $qr_count = (int)$request->qr_count - $career->qr_count;
+                $counter = 1;
                 while ($qr_count) {
                     $random = Str::random(10);
                     $link = "famenu.ir/qrcodes/$career->id/" . $random;
@@ -144,9 +251,11 @@ class CareerController extends Controller
                         'career_id' => $career->id,
                         'page_path' => 'qrcode/' . $career->id . '/' . $random,
                         'slug' => $random,
+                        'description'=>'میز '.$counter,
                         'user_id' => Auth::id()
                     ]);
                     $qr_count--;
+                    $counter++;
                 }
             }
         }
@@ -170,25 +279,44 @@ class CareerController extends Controller
                             foreach ($category->menu_items as $item) {
                                 if (count($item->ingredients)) {
                                     foreach ($item->ingredients as $ingredients) {
+                                         if ($ingredients->image) {
+                                        Storage::disk('public')->delete($ingredients->image);
+                                        }
                                         $ingredients->delete();
                                     }
+                                }
+                                 if ($item->image) {
+                                Storage::disk('public')->delete($item->image);
                                 }
                                 $item->delete();
                             }
                         }
+                        if ($category->image) {
+                        Storage::disk('public')->delete($category->image);
+                        }
                         $category->delete();
                     }
+                }
+                 if ($menu->banner) {
+                Storage::disk('public')->delete($menu->banner);
                 }
                 $menu->delete();
             }
         }
+         if ($career->logo) {
+          Storage::disk('public')->delete($career->logo);
+            }
+         if ($career->banner) {
+          Storage::disk('public')->delete($career->banner);
+            }
         $career->delete();
-        return to_route('career.careers');
+        return redirect()->back();
     }
 
     public function index()
     {
         $careers = career::all();
+
         return view('admin.careers.index', ['careers' => $careers]);
     }
 
@@ -199,7 +327,10 @@ class CareerController extends Controller
 
     public function menus(career $career)
     {
-        return view('admin.careers.menuList', ['career' => $career]);
+//         return view('admin.careers.menuList', ['career' => $career]);
+        $menus = $career->menus;
+        return response()->json($menus);
+        
     }
 
     public function qr_codes(career $career)
@@ -227,67 +358,92 @@ class CareerController extends Controller
                                 foreach ($category->menu_items as $item) {
                                     if (count($item->ingredients)) {
                                         foreach ($item->ingredients as $ingredients) {
+                                              if ($ingredients->image) {
+                                                Storage::disk('public')->delete($ingredients->image);
+                                                }
                                             $ingredients->delete();
                                         }
+                                    }
+                                    if ($item->image) {
+                                    Storage::disk('public')->delete($item->image);
                                     }
                                     $item->delete();
                                 }
                             }
+                             if ($category->image) {
+                             Storage::disk('public')->delete($category->image);
+                            }
                             $category->delete();
                         }
                     }
+                     if ($menu->banner) {
+                     Storage::disk('public')->delete($menu->banner);
+                    }
                     $menu->delete();
                 }
+            }
+        if ($career->logo) {
+          Storage::disk('public')->delete($career->logo);
+            }
+         if ($career->banner) {
+          Storage::disk('public')->delete($career->banner);
             }
             $career->delete();
         }
         return redirect()->back();
     }
+
     public function careersCategories()
     {
         $careerCategories = careerCategory::all();
-        return view('client.career.careersCategories' , ['careerCategories'=> $careerCategories]);
+        return view('client.career.careersCategories', ['careerCategories' => $careerCategories]);
     }
+
     public function careersList()
     {
         $careers = career::all();
-        return view('client.career.careersList' , ['careers'=> $careers]);
+        return view('client.career.careersList', ['careers' => $careers]);
     }
+
     public function categoryCareers(careerCategory $careerCategory)
     {
         return view('client.career.categoryCareers', ['careerCategory' => $careerCategory]);
     }
 
-    public function orders(career $career){
+    public function orders(career $career)
+    {
 
-        return view('admin.careers.orders', ['career'=>$career]);
+        return view('admin.careers.orders', ['career' => $career]);
     }
 
-    public function acceptOrder(Request $request){
+    public function acceptOrder(Request $request)
+    {
         $order = order::find($request->order_id);
-        $request->state=='accept' && $order->status = 2;
-        $request->state=='send' && $order->status = 3;
+        $request->state == 'accept' && $order->status = 2;
+        $request->state == 'send' && $order->status = 3;
 
         $order->save();
         return response()->json($order);
     }
 
-    public function createUser(){
+    public function createUser()
+    {
         return view('admin.careers.createUser');
     }
 
-    public function storeUser(Request $request){
+    public function storeUser(Request $request)
+    {
         $password = Hash::make($request->password);
         $user_id = User::insertGetId([
-            'name'=>$request->name,
-            'family'=>$request->family,
-            'phoneNumber'=>$request->phoneNumber,
-            'password'=>$password,
-            'parent_id'=>Auth::id()
+            'name' => $request->name,
+            'family' => $request->family,
+            'phoneNumber' => $request->phoneNumber,
+            'password' => $password,
+            'parent_id' => Auth::id()
         ]);
         role_user::create([
-            'role_id'=>3,
-            'user_id'=>$user_id
+            'role_id' => 3,
+            'user_id' => $user_id
         ]);
         return to_route('career.create', [$user_id]);
     }
