+++ app/Http/Controllers/CartController.php
@@ -4,39 +4,62 @@ namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use App\Models\cart;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Log;
 
 class CartController extends Controller
 {
     public function store(Request $request)
     {
+//        return response()->json($request->all());
+        $user_id = Auth::id();
+        if (!Auth::check()) {
+            $user_id = $request->input('user_id');
+        }
+
         $cart = cart::create([
             'career_id' => $request->career_id,
             'menu_item_id' => $request->menu_item_id,
-            'user_id' => Auth::id(),
+            'user_id' => $user_id,
             'quantity' => $request->quantity ? $request->quantity : 1,
         ]);
+//        dd($cart);
         return response()->json($cart);
     }
 
-    function update(Request $request){
-        $cart = cart::where(['career_id'=>$request->career_id, 'menu_item_id'=>$request->menu_item_id, 'user_id'=>Auth::id(), 'order_id'=>null])->first();
-        $cart->quantity = $request->quantity;
+    function update(Request $request)
+    {
+        $user_id = $request->input('user_id');
+        if (Auth::check()) {
+            $user_id = Auth::id();
+        }
+        $cart = cart::where(['career_id' => $request->career_id, 'menu_item_id' => $request->menu_item_id, 'user_id' => $user_id, 'order_id' => null])->first();
+        $cart->quantity = $request->quantity ? $request->quantity : 1;
         $cart->save();
         return response()->json([$cart, $cart->quantity, $request->quantity]);
     }
 
-    public function showcarts(Request $request)
+    public function showCarts(Request $request)
     {
-        $data = Auth::user()->load(['carts'=>function($query){
+        $user = null;
+        if (isset($request->user_id)) {
+            $user = User::find($request->user_id);
+        }
+        if (Auth::check()) {
+            $user = Auth::user();
+        }
+        $data = $user->load(['carts' => function ($query) {
             $query->whereNull('order_id')->with('menu_item')->get();
         }]);
+//        return response()->json($data);
         $datas = [];
         foreach ($data->carts as $cart) {
+            Log::info($cart);
             $cart->menu_item['quantity'] = $cart->quantity;
             $cart->menu_item['cart_id'] = $cart->id;
-            $cart->menu_item['quantity'] = $cart->quantity;
+//            $cart->menu_item['quantity'] = $cart->quantity;
             $datas[] = $cart->menu_item;
         }
         return response()->json($datas);
@@ -52,9 +75,9 @@ class CartController extends Controller
         return response()->json($cart);
     }
 
-    public function delete(Request $request){
-        $cart = cart::where(['menu_item_id'=>$request->menu_item_id, 'user_id'=>Auth::id()])->first();
-        $cart->delete();
-        return response()->json('ok');
+    public function delete(Request $request)
+    {
+        $cart = cart::where('menu_item_id', $request->menu_item_id)->where('user_id', $request->user_id)->delete();
+        return response()->json($cart);
     }
 }
